+++ scripts/ai-bridge/internal/usecase/diagnose_test.go
@@ -8,7 +8,7 @@ import (
 
 	"ai-bridge/internal/domain"
 	"ai-bridge/internal/usecase"
-	"ai-bridge/internal/usecase/mock"
+	"ai-bridge/internal/usecase/port/mock"
 )
 
 func TestDiagnoseRun(t *testing.T) {
