+++ scripts/ai-bridge/internal/usecase/port/port.go
@@ -1,11 +1,13 @@
-// Package usecase holds the application business rules of ai-bridge.
+// Package port declares the interfaces (ports) that the use case layer depends
+// on for layer-crossing side effects.
 //
-// It orchestrates the domain layer and depends only on domain types and the
-// ports declared in this file. Ports are interfaces owned by this (inner) layer
-// and implemented by the outer infrastructure layer, inverting the dependency
-// so that side effects (filesystem, process, environment) can be mocked in
-// tests. Mocks are generated by mockgen via the directive below.
-package usecase
+// Ports are owned by the inner application layer and implemented by the outer
+// infrastructure layer, inverting the dependency so that side effects
+// (filesystem, process, environment) can be mocked in tests. Naming a call site
+// argument or struct field with a port.* type makes it explicit that the call
+// crosses a boundary rather than invoking the use case's own method. Mocks are
+// generated by mockgen via the directive below.
+package port
 
 import (
 	"context"
