+++ scripts/ai-bridge/internal/usecase/run_daemon_test.go
@@ -9,7 +9,7 @@ import (
 
 	"ai-bridge/internal/domain"
 	"ai-bridge/internal/usecase"
-	"ai-bridge/internal/usecase/mock"
+	"ai-bridge/internal/usecase/port/mock"
 )
 
 func TestRunDaemonRun(t *testing.T) {
